+++ Positive or negative/Program.cs
@@ -11,7 +11,7 @@ namespace Positive_or_negative
         static void Main(string[] args)
         {
             Console.Write("Enter number: ");
-            int a = Convert.ToInt32(Console.ReadLine());
+            double a = Convert.ToDouble(Console.ReadLine());
             if (a >= 0)
                 Console.WriteLine($"{a} is a positive number");
             else Console.WriteLine($"{a} is a negative number");
